+++ bungeecord/src/main/java/me/neznamy/tab/platforms/bungeecord/BungeeTabPlayer.java
@@ -76,7 +76,7 @@ public class BungeeTabPlayer extends ProxyTabPlayer {
 	}
 	
 	@Override
-	public long getPing() {
+	public int getPing() {
 		return player.getPing();
 	}
