+++ shared/src/main/java/me/neznamy/tab/shared/features/GlobalPlayerlist.java
@@ -159,7 +159,7 @@ public class GlobalPlayerlist extends TabFeature {
 			format = playerlist.getTabFormat(p, viewer);
 		}
 		return new PacketPlayOutPlayerInfo(EnumPlayerInfoAction.ADD_PLAYER, new PlayerInfoData(p.getName(), p.getTablistUUID(), p.getSkin(), 
-				(int)p.getPing(), isVanishedAsSpectators() && p.isVanished() ? EnumGamemode.SPECTATOR : EnumGamemode.CREATIVE, format));
+				p.getPing(), isVanishedAsSpectators() && p.isVanished() ? EnumGamemode.SPECTATOR : EnumGamemode.CREATIVE, format));
 	}
 
 	@Override
